Minor PEP8 changes
