+++ aimmo-game/simulation/game_logic/spawn_location_finder.py
@@ -14,10 +14,10 @@ class SpawnLocationFinder:
         Used to make sure that the cell is free before spawning.
         """
         return (c for c in self._world_map.all_cells()
-                if c.habitable
-                and not c.generates_score
-                and not c.avatar
-                and not c.pickup)
+                if c.habitable and not
+                c.generates_score and not
+                c.avatar and not
+                c.pickup)
 
     def get_random_spawn_locations(self, max_locations):
         if max_locations <= 0:
