+++ aimmo-game/simulation/map_generator.py
@@ -44,8 +44,8 @@ class Main(_BaseGenerator):
         always_empty_location = Location(always_empty_edge_x, always_empty_edge_y)
 
         for cell in shuffled(world_map.all_cells()):
-            if (cell.location != always_empty_location
-                    and random.random() < self.settings['OBSTACLE_RATIO']):
+            if (cell.location != always_empty_location and
+                    random.random() < self.settings['OBSTACLE_RATIO']):
                 cell.habitable = False
                 # So long as all habitable neighbours can still reach each other, then the
                 # map cannot get bisected.
