+++ aimmo-game/simulation/worker_managers/__init__.py
@@ -2,6 +2,6 @@ from kubernetes_worker_manager import KubernetesWorkerManager as _KWorkerManager
 from local_worker_manager import LocalWorkerManager as _LWorkerManager
 
 WORKER_MANAGERS = {
-    'local': _KWorkerManager,
-    'kubernetes': _LWorkerManager,
+    'local': _LWorkerManager,
+    'kubernetes': _KWorkerManager,
 }
