+++ aimmo-game/tests/functional/test_invulnerability_pickups_and_effects.py
@@ -41,7 +41,7 @@ class TestInvulnerabilityPickupsAndEffects(TestCase):
         self.assertEqual(self.cell.avatar, self.game.avatar_manager.get_avatar(1))
         self.assertEqual(self.cell.avatar.resistance, 1000)
 
-        self.cell = self.game.game_state.world_map.get_cell(Location(2,0))
+        self.cell = self.game.game_state.world_map.get_cell(Location(2, 0))
         self.cell.pickup = InvulnerabilityPickup(self.cell)
         self.game.turn_manager._run_single_turn()
