+++ aimmo-game/tests/test_simulation/avatar/test_avatar_wrapper.py
@@ -8,7 +8,6 @@ from simulation.avatar import avatar_wrapper
 from simulation.location import Location
 
 
-
 class MockEffect(object):
     def __init__(self, avatar):
         self.turns = 0
